all Spam tests pass with 1M timeout and 2K max script size

+++ src/main/scala/sigmastate/eval/Evaluation.scala
@@ -469,7 +469,7 @@ trait Evaluation extends RuntimeCosting { IR: IRContext =>
     case _ => error(s"Cannot find value in environment for $s (dataEnv = $dataEnv)")
   }
 
-  def msgCostLimitError(cost: Int, limit: Long) = s"Estimated expression complexity $cost exceeds the limit $limit"
+  def msgCostLimitError(cost: Long, limit: Long) = s"Estimated expression complexity $cost exceeds the limit $limit"
 
   /** Incapsulate simple monotonic (add only) counter with reset. */
   class CostCounter(val initialCost: Int) {
